+++ conduit/Data/Conduit/List.hs
@@ -1,6 +1,7 @@
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE CPP #-}
+{-# LANGUAGE Trustworthy #-}
 -- | Higher-level functions to interact with the elements of a stream. Most of
 -- these are based on list functions.
 --
@@ -510,7 +511,7 @@ concatMapMC f = awaitForever $ sourceList <=< lift . f
 {-# INLINE concatMapMC #-}
 STREAMING(concatMapM, concatMapMC, concatMapMS, f)
 
--- | 'concatMap' with an accumulator.
+-- | 'concatMap' with a strict accumulator.
 --
 -- Subject to fusion
 --
@@ -535,7 +536,8 @@ scanlM f s = void $ mapAccumM f s
 {-# DEPRECATED scanlM "Use mapAccumM instead" #-}
 
 -- | Analog of @mapAccumL@ for lists. Note that in contrast to @mapAccumL@, the function argument
---   takes the accumulator as its second argument, not its first argument.
+--   takes the accumulator as its second argument, not its first argument, and the accumulated value
+--   is strict.
 --
 -- Subject to fusion
 --
@@ -544,7 +546,7 @@ mapAccum, mapAccumC :: Monad m => (a -> s -> (s, b)) -> s -> ConduitM a b m s
 mapAccumC f =
     loop
   where
-    loop s = await >>= maybe (return s) go
+    loop !s = await >>= maybe (return s) go
       where
         go a = case f a s of
                  (s', b) -> yield b >> loop s'
@@ -559,7 +561,7 @@ mapAccumM, mapAccumMC :: Monad m => (a -> s -> m (s, b)) -> s -> ConduitM a b m 
 mapAccumMC f =
     loop
   where
-    loop s = await >>= maybe (return s) go
+    loop !s = await >>= maybe (return s) go
       where
         go a = do (s', b) <- lift $ f a s
                   yield b
@@ -583,7 +585,7 @@ INLINE_RULE(scan, f, mapAccum (\a b -> let r = f a b in (r, r)))
 scanM :: Monad m => (a -> b -> m b) -> b -> ConduitM a b m b
 INLINE_RULE(scanM, f, mapAccumM (\a b -> f a b >>= \r -> return (r, r)))
 
--- | 'concatMapM' with an accumulator.
+-- | 'concatMapM' with a strict accumulator.
 --
 -- Subject to fusion
 --