+++ network-conduit-tls/ChangeLog.md
@@ -1,3 +1,11 @@
+## 1.2.2
+
+* Make runTLS{Client,Server}StartTLS general [#264](https://github.com/snoyberg/conduit/pull/264)
+
+## 1.2.1
+
+* Expose ApplicationStartTLS [#262](https://github.com/snoyberg/conduit/pull/262)
+
 ## 1.2.0.1
 
 * tls 1.3 support