+++ resourcet/ChangeLog.md
@@ -1,3 +1,36 @@
+## 1.1.8
+
+* Add `instance MonadFix ResourceT`
+  [#281](https://github.com/snoyberg/conduit/pull/281)
+
+## 1.1.7.5
+
+* Inline the tutorial from SoH
+
+## 1.1.7.4
+
+* Make test suite slightly more robust
+
+## 1.1.7.3
+
+* Doc tweak
+
+## 1.1.7.2
+
+* Remove upper bound on transformers [#249](https://github.com/snoyberg/conduit/issues/249)
+
+## 1.1.7.1
+
+* transformers-compat 0.5
+
+## 1.1.7
+
+* Canonicalise Monad instances [#237](https://github.com/snoyberg/conduit/pull/237)
+
+## 1.1.6
+
+* Safe/Trustworthy for resourcet [#220](https://github.com/snoyberg/conduit/pull/220)
+
 ## 1.1.5
 
 *  Add pass-through instances for Alternative and MonadPlus [#214](https://github.com/snoyberg/conduit/pull/214)